+++ lib/providers/factories/ollama_factory.dart
@@ -1,7 +1,7 @@
 import '../../core/capability.dart';
 import '../../core/config.dart';
-import '../../src/config/provider_defaults.dart';
 import '../../src/compatibility/providers/ollama/config_adapter.dart';
+import '../ollama/defaults.dart';
 import '../ollama/ollama.dart';
 import 'base_factory.dart';
 
@@ -37,6 +37,9 @@ class OllamaProviderFactory extends LocalProviderFactory<ChatCapability> {
 
   @override
   Map<String, dynamic> getProviderDefaults() {
-    return ProviderDefaults.getDefaults('ollama');
+    return {
+      'baseUrl': OllamaDefaults.baseUrl,
+      'model': OllamaDefaults.defaultModel,
+    };
   }
 }